refactor: 검색 결과 관련 컴포넌트들 별도의 폴더로 구성
