+++ src/components/search-result/AlarmModal.tsx
