+++ src/components/search-result/KoreanStandardTime.tsx
